Style and dataApprovalLevel updates
Switched to using an appstyle for theming to be able to
style app specific components.
Fixed dataApprovalLevels to be work against the reworked api for
adding levels.
Changed the sidebar to be more in line with the currents system
style so the new app doesn't fall out of place too much.